+++ src/e2e/app/cli/repo-tests/src/commands/test_add_command.rs
@@ -254,3 +254,46 @@ pub async fn test_add_with_circular_dependency(kamu: KamuCliPuppet) {
 }
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+
+pub async fn test_add_with_duplicate_dependencies(kamu: KamuCliPuppet) {
+    let snapshot = MetadataFactory::dataset_snapshot().name("plain").build();
+    let manifest = odf::serde::yaml::YamlDatasetSnapshotSerializer
+        .write_manifest_str(&snapshot)
+        .unwrap();
+    std::fs::write(kamu.workspace_path().join("plain.yaml"), manifest).unwrap();
+
+    let snapshot = MetadataFactory::dataset_snapshot()
+        .name("with-duplicates")
+        .kind(odf::DatasetKind::Derivative)
+        .push_event(
+            MetadataFactory::set_transform()
+                .inputs_from_refs(["plain", "plain"])
+                .build(),
+        )
+        .build();
+    let manifest = odf::serde::yaml::YamlDatasetSnapshotSerializer
+        .write_manifest_str(&snapshot)
+        .unwrap();
+    std::fs::write(kamu.workspace_path().join("with-duplicates.yaml"), manifest).unwrap();
+
+    kamu.assert_failure_command_execution(
+        ["-v", "add", "plain.yaml", "with-duplicates.yaml"],
+        None,
+        Some([
+            "Failed to add dataset with-duplicates: Invalid snapshot: Invalid event: Transform \
+             contains duplicate inputs",
+        ]),
+    )
+    .await;
+
+    let dataset_names = kamu
+        .list_datasets()
+        .await
+        .into_iter()
+        .map(|dataset| dataset.name)
+        .collect::<Vec<_>>();
+
+    assert!(dataset_names.is_empty());
+}
+
+////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////